// src/interfaces/OrderDetail.ts
export type OrderDetail = {
  OrderDetailId: number;
  CustomerName: string;
  CustomerEmail: string;
  CustomerAddress: string;
  CustomerDni: string;
  CustomerRegistrationDate: Date;
  ProductName: string;
  ProductDescription: string;
  ProductPrice: number;
  ProductQuantity: number;
  TotalPrice: number;
};